changed coord types to double

+++ include/mission_definition_file.hpp
@@ -17,7 +17,7 @@ namespace mission_file_lib
      */
     struct safety
     {
-        std::vector<std::pair<float, float>> geofence_points;                  ///< Vector of geofence points. Each point is represented by a pair of floats (latitude, longitude).
+        std::vector<std::pair<double, double>> geofence_points;                  ///< Vector of geofence points. Each point is represented by a pair of floats (latitude, longitude).
         uint16_t max_height_cm = common_lib::MAX_FLIGHT_HEIGHT_CM;             ///< Maximum flight height in centimeters.
         uint16_t min_cruise_height_cm = common_lib::MIN_CRUISE_HEIGHT_CM;      ///< Minimum cruise height in centimeters.
         float max_horizontal_speed_mps = common_lib::MAX_HORIZONTAL_SPEED_MPS; ///< Maximum horizontal speed in meters per second.
